PythonProject/requirements.txt
fastapi
uvicorn[standard]
groq
sentence-transformers
faiss-cpu
numpy
python-dotenv
